Fix Float casting

+++ MMOCoreORB/src/server/zone/managers/combat/CombatManager.cpp
@@ -931,20 +931,20 @@ float CombatManager::applyDamageModifiers(CreatureObject* attacker, WeaponObject
 	}
 	if (data.isForceAttack()){
 		ManagedReference<PlayerObject*> attackerGhost = attacker->getPlayerObject();
-		int forcePower = 0;
+		float forcePower = 0.f;
 		if(attackerGhost->getJediState() == 4) {
 			info("Light Force Power Jedi : " + String::valueOf(attacker->getSkillMod("force_power_light")), true );
-			forcePower = attacker->getSkillMod("force_power_light")/100;
+			forcePower = (float)attacker->getSkillMod("force_power_light") / 100;
 	} 	else if (attackerGhost->getJediState() == 8) {
 			info("Dark Force Power Jedi : " + String::valueOf(attacker->getSkillMod("force_power_dark")), true );
-			forcePower = attacker->getSkillMod("force_power_dark")/100;
+			forcePower = (float)attacker->getSkillMod("force_power_dark") / 100;
 	}
 	
-	info("Force power increase is: " + String::valueOf(forcePower), true);
-	info("Old Damage: " + String::valueOf(damage), true);
-	int frsDamage = damage * forcePower;
-	damage += frsDamage;
-	info("New Damage: " + String::valueOf(damage), true);
+		info("Force power increase is: " + String::valueOf(forcePower), true);
+		info("Old Damage: " + String::valueOf(damage), true);
+		int frsDamage = damage * forcePower;
+		damage += frsDamage;
+		info("New Damage: " + String::valueOf(damage), true);
 	}
 
 	damage += attacker->getSkillMod("private_damage_bonus");
